Removing Blog
Since the active blog is on the club site, and we don't need two
at the moment, removing the link to this blog on the org page.

Note: The blog will still exist, there just won't be a link to
it anywhere visible. Easy to re-add a link if/when we decide to
start updating that again.